requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
certifi==2023.7.22
charset-normalizer==3.3.0
chroma-hnswlib==0.7.3
chromadb==0.4.14
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.1
fastapi==0.103.2
filelock==3.12.4
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.2
grpcio==1.59.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-resources==6.1.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.312
langsmith==0.0.43
marshmallow==3.20.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.0
onnxruntime==1.16.0
openai==0.28.1
overrides==7.4.0
packaging==23.2
posthog==3.0.2
protobuf==4.24.4
pulsar-client==3.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pypdf==3.16.4
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.1
tokenizers==0.14.1
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
